Block ids past 255: u16 chunk storage + RLE v2 with legacy decode
Chunk.data is now a Uint16Array (BlockData) end to end: generation,
gen/mesh worker transfers, the mesher's padded copy and per-id LUTs
(sized Blocks.ID_LIMIT = 4096). Chunk blobs are saved as RLE v2
([0,0,2] header + [count u16, id u16] runs); headerless v1 u8 blobs
from older saves still decode. Blocks may now use 256-299 and 1000+.

Node bench (25 chunks): gen 2.7->2.9 ms overworld, nether ~1.4 ms,
mesh 3.0->3.1 ms, RLE 8.1->10.9 KB per modified chunk, chunk RAM
40->80 KB.

+++ src/engine/NetherFX.ts
@@ -103,7 +103,7 @@ const rnd = (a: number, b: number): number => a + Math.random() * (b - a);
 
 interface FXWorld {
   getBlock(x: number, y: number, z: number): number;
-  getChunk(cx: number, cz: number): { cx: number; cz: number; ready: boolean; data: Uint8Array; glowers: Set<number> } | undefined;
+  getChunk(cx: number, cz: number): { cx: number; cz: number; ready: boolean; data: Uint16Array; glowers: Set<number> } | undefined;
 }
 
 export class NetherFX {
